Do not assume that namespaces start with `http://opcfoundation.org/UA/`
The function `NodeSetToModelDesign.CreateNamespace` previously assumed that all namespaces started with `http://opcfoundation.org/UA/` and removed a substring of this length from the start of the namespace. This lead to a crash if the namespace did not start with this prefix and was in total shorter.

Check if the namespace starts with this prefix, before removing it.

If this prefix is not present, then check for the shorter prefix `http://` and remove that.

Then continue with the existing algorithm to replace slashes and spaces.